Fixed the makefile
