+++ Makefile
@@ -21,7 +21,7 @@ DEPEXT      := d
 OBJEXT      := o
 
 #Flags, Libraries and Includes
-CXXFLAGS    := -std=c++98
+CXXFLAGS    := -std=c++98 -Wall -Werror -Wextra -pedantic -std=c++98
 LIB         := -lncurses
 INC         := -I$(INCDIR) -I/usr/local/include
 INCDEP      := -I$(INCDIR)
